requirements.txt
absl-py==2.3.1
attrs==25.3.0
cffi==1.17.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.2
jax==0.7.1
jaxlib==0.7.1
kiwisolver==1.4.9
matplotlib==3.10.5
mediapipe==0.10.18
minimalmodbus==2.1.1
ml_dtypes==0.5.3
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
protobuf==4.25.8
pycparser==2.22
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
scipy==1.16.1
sentencepiece==0.2.1
six==1.17.0
sounddevice==0.5.2
